pyproject.toml
[project]
name = "findata-library"
version = "0.0.1"
dependencies = [
    'boto3==1.35.30',
    'psycopg2-binary==2.9.9',
    'polygon-api-client==1.14.2',
    'aiohttp==3.10.5',
    'asyncio-throttle==1.0.2',
    'pandas==2.2.3',
    'asyncpg==0.30.0'
]

[tool.setuptools.packages.find]
include = ["ontrade_lib"]
